+++ dart/lib/src/protocol/debug_meta.dart
@@ -18,6 +18,20 @@ class DebugMeta {
 
   DebugMeta({this.sdk, List<DebugImage>? images}) : _images = images;
 
+  /// Deserializes a [DebugMeta] from JSON [Map].
+  factory DebugMeta.fromJson(Map<String, dynamic> json) {
+    final sdkInfoJson = json['sdk_info'];
+    final debugImagesJson = json['images'] as List<dynamic>?;
+    return DebugMeta(
+      sdk: sdkInfoJson != null ? SdkInfo.fromJson(sdkInfoJson) : null,
+      images: debugImagesJson
+          ?.map((debugImageJson) =>
+              DebugImage.fromJson(debugImageJson as Map<String, dynamic>))
+          .toList(),
+    );
+  }
+
+  /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
     final json = <String, dynamic>{};