+++ dart/lib/src/transport/noop_transport.dart
@@ -1,9 +1,12 @@
 import 'dart:async';
 
+import '../sentry_envelope.dart';
+
 import '../protocol.dart';
 import 'transport.dart';
 
 class NoOpTransport implements Transport {
   @override
-  Future<SentryId> send(SentryEvent event) => Future.value(SentryId.empty());
+  Future<SentryId> send(SentryEnvelope envelope) =>
+      Future.value(SentryId.empty());
 }