+++ dart/test/mocks/mock_transport.dart
@@ -1,15 +1,15 @@
 import 'package:sentry/sentry.dart';
 
 class MockTransport implements Transport {
-  List<SentryEvent> events = [];
+  List<SentryEnvelope> envelopes = [];
 
   bool called(int calls) {
-    return events.length == calls;
+    return envelopes.length == calls;
   }
 
   @override
-  Future<SentryId> send(SentryEvent event) async {
-    events.add(event);
-    return event.eventId;
+  Future<SentryId> send(SentryEnvelope envelope) async {
+    envelopes.add(envelope);
+    return envelope.header.eventId ?? SentryId.empty();
   }
 }