+++ flutter/test/load_android_image_list_test.dart
@@ -113,7 +113,7 @@ void main() {
 
     final ep = fixture.options.eventProcessors.first;
     SentryEvent? event = getEvent();
-    event = await ep(event);
+    event = await ep.apply(event);
 
     expect(1, event!.debugMeta!.images.length);
   });
@@ -124,7 +124,7 @@ void main() {
     sut.call(fixture.hub, fixture.options);
     final ep = fixture.options.eventProcessors.first;
     SentryEvent? event = getEvent();
-    event = await ep(event);
+    event = await ep.apply(event);
 
     final image = event!.debugMeta!.images.first;